[ PrimeFactorization.h Changes ]
    - changed ull typedef to be based on uint64_t define from the <cstdint>, for consistency
[ RSA.cpp Changes ]
    - changed the CalculateModularPow to use the Right-to-left binary method defined in the wiki, huge speed increase for decryption

+++ EncryptionTechnique/EncryptionTechnique/PrimeFactorization.h
@@ -2,10 +2,11 @@
 #define _PRIME_FACTORIZATION_H_INCLUDED_
 
 #include <vector>
+#include <cstdint>
 
 namespace EncryptionSequence
 {
-  typedef unsigned long long ull;
+  typedef uint64_t ull;
 
   class PrimeFactor
   {
